refactor: :building_construction: update bot file structure (#59)
