+++ bot/metadata/index.js
@@ -5,8 +5,8 @@ import {
   gatherRepoAuthors,
   getDOI,
   createId,
-} from "../tools/index.js";
-import dbInstance from "../../db.js";
+} from "../utils/tools/index.js";
+import dbInstance from "../db.js";
 
 const CODEFAIR_DOMAIN = process.env.CODEFAIR_APP_DOMAIN;
