+++ bot/utils/renderer/index.js
@@ -4,9 +4,9 @@ import {
   createId,
   applyLastModifiedTemplate,
 } from "../tools/index.js";
-import { applyCWLTemplate } from "../cwl/index.js";
-import { applyMetadataTemplate } from "../metadata/index.js";
-import { applyLicenseTemplate } from "../license/index.js";
+import { applyCWLTemplate } from "../../cwl/index.js";
+import { applyMetadataTemplate } from "../../metadata/index.js";
+import { applyLicenseTemplate } from "../../license/index.js";
 
 const { GITHUB_APP_NAME } = process.env;
 const CODEFAIR_DOMAIN = process.env.CODEFAIR_APP_DOMAIN;
